(1) high_break needs to be cast to float64 (or something higher) as numpy doesnt anymore and it causes nans; and (2)index->indx again for the warm target model to keep it unique from the cold thick target inputs

+++ sunkit_spex/legacy/fitting/photon_models_for_fitting.py
@@ -88,7 +88,7 @@ def thick_fn(total_eflux, index, e_c, energies=None):
     # we want a single power law electron distribution,
     # so set eebrk == eehigh at a high value.
     # we don't care about q at E > eebrk.
-    high_break = energies.max() * 10
+    high_break = np.float64(energies.max() * 10)
     output = bremsstrahlung_thick_target(
         photon_energies=energies, p=index, eebrk=high_break, q=20, eelow=e_c, eehigh=high_break
     )
